+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.1.0
+- <cj>Snap mode</c> replaced with <cg>No-anchor mode</c> and partially with <cg>Grid-snap mode</c>
+- <cj>Free rotation button</c> replaced with <cg>additional transform element</c>
+- Added <cg>Grid-snap mode</c>
+
 # 1.0.1
 - fix RobTop's crash with extremely thin objects (thanks, Brittank88)